Move prompt code to diffrent class
